Test flit_core.sdist
